Tokens TypeScript Build (#31)
# Pull Request

## 🤨 Rationale

Added a workaround that avoids forking the DSP generated config.js and enables TypeScript Type Definitions for the JS Build. Could remove after: https://github.com/AdobeXD/design-system-package-dsp/issues/22

## 👩‍💻 Implementation

Specified a custom config file in the DSP configuration. The custom file loads the one generated by the design system and patches it for the TypeScript configuration changes.

I noticed that even with a custom file specified the DSP plugin will still assume the original config.js exists and will always create or update it. The extend and patch workflow in `nimble-config.js` should allow us to leverage the GUI config for the DSP to generate a `config.js` and patch it for custom behavior in `nimble-config.js`.

## 🧪 Testing

Tested with the XD extension locally.

## ✅ Checklist

- [X] I have updated the project documentation to reflect my changes. Don't think any updates needed.