Switch to ES6 builds
Using ESNext ends up creating classes with public class fields. For button that looks like `class Button { appearance; }`. The class property initializer ends up assigning a property to Button instances. This causes the getter and setter from the @attr on the prototype chain to be ignored and the owned property is assigned instead.