Add nimble-checkbox component (#58)
🤨 Rationale

Create a <nimble-checkbox> component according to this design spec.

👩‍💻 Implementation

I re-used the Checkbox component from @microsoft/fast-foundation and applied some styling to it. The foundation component expects a checked indicator in form of an HTML element. I re-used the SVG from an existing checked icon from nimble tokens.

I created stories for an enabled and disabled checkbox with controls for label, checked, and disabled parameters. The stories are implemented using a litte different approach from other components in this package. This approach aligns better with the current documentation and enables better type checking.

🧪 Testing

I tested the component in Chrome, Firefox, and Edge and ensured that it aligns with the design spec. Unit tests are not needed because we re-use the foundation component without any additional logic.

✅ Checklist

 X - I have updated the project documentation to reflect my changes or determined no changes are needed.