Component test matrices (#76)
🤨 Rationale

We are currently using Chromatic to test only the light theme, so our visual test coverage is poor. This endeavors to fix that.

👩‍💻 Implementation

Created a new Story for each component that creates matrix of component themes and states. E.g.

Screen Shot 2021-08-16 at 2 41 19 PM

Also, deleted some of the stories that were only being used to exercise the state that now captured in the matrix story for each component.

🧪 Testing

Local build and manual Storybook testing

✅ Checklist

 I have updated the project documentation to reflect my changes or determined no changes are needed.