Storybook matrix states (#88)
# Pull Request

## 🤨 Rationale

The goal of this PR is to make managing the matrix of states for storybook tests more maintainable. The matrixThemeHelper was modified to take an array of states. The helper calculates all the permutations of those states and applies them to the provided component template.

## 👩‍💻 Implementation

Updated the `matrixThemeHelper ` and current usages. 

## 🧪 Testing

Built the storybook locally. Updates should be reviewed in PR.

## ✅ Checklist

- [X] I have updated the project documentation to reflect my changes or determined no changes are needed.