Text field invalid state support (#86)
## 🤨 Rationale

This PR adds a new boolean attribute (and the appropriate styling) for the invalid state for the text-field component, per [the latest designs](https://xd.adobe.com/view/8ce280ab-1559-4961-945c-182955c7780b-d9b1/screen/842889a5-67ba-4350-91c1-55eee48f4fa2/specs/). Note that I opted not to extend these changes to the number-field because the number-field design is still in flux and the positioning of the +/- glyphs may need to be adjusted.

Next step here is to extend this new state support to the Angular wrapper for the text-field, but I will make that a separate PR.

## 👩‍💻 Implementation

- Added a new `invalid` boolean attribute to the text-field component (which required extending `FoundationTextField`). This class also sets the slot content to the invalid icon.
    - Note that we can't set the slot content as we have for other components because of the way the available slots on the text-field are defined. I asked about this on the [discord channel](https://discord.com/channels/449251231706251264/482296787021856768/880214188276985906) but I haven't gotten an answer.
- New styling for the invalid state as well as ensuring that the styling is structured such that the disabled state takes precedence if both states are true.
- Added toggles in the primary text-field storybook story to allow toggling of both the disabled and invalid states (I also added a disabled state toggle to the number-field)


## 🧪 Testing

Storybook/local testing.