Storybook component doc templates (#96)
# Pull Request

## 🤨 

The Design System components should be fully documented - both from an API and a design usage perspective. This change makes it possible for visual and interaction designers to modify simple Markdown-like files (*.mdx) and have that documentation show up in Storybook next to the components. 

![Screen Shot 2021-09-03 at 9 46 42 AM](https://user-images.githubusercontent.com/1458528/132025516-0e9f7b20-cf90-4e0e-bf62-1ec0000994ed.png)


## 👩‍💻 Implementation

This change makes it possible to add a `*-docs.stories.mdx` to a component `tests` folder, and keeps the documentation page separate from the core component hosting story.  
Also, updated the docs, and created a `CONTRIBUTING-DOCS.md` to document this process.

## 🧪 Testing

Manual storybook review

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.