Style password reveal in Edge (#99)
## 🤨 Rationale

This change is intended to address #97 . 

## 👩‍💻 Implementation

Edge adds a password reveal icon to inputs with type="password" (docs [here](https://docs.microsoft.com/en-us/microsoft-edge/web-platform/password-reveal)). This change styles it to match the color of the text in each of the themes.

## 🧪 Testing

Tested in Edge to confirm that the icon is colored correctly. This change should be inert for other browsers.