Style text-fields in matrix for visual readability  (#106)
## 🤨 Rationale

As outlined in #103, we've had some challenges with the visual diffs of the text-field matrix as it has grown (and it is growing further in #105). This felt like a quick and dirty way to dramatically improve the visual readability of the matrix without impacting the breadth/simplicity of our visual tests.

## 👩‍💻 Implementation

Imposed a fixed width and padding for text-fields in the matrix.