Upgrade fast-foundation and leverage end slot support for text-field (#108)
## 🤨 Rationale

When we first added the invalid state styling for the text-field, the end slot did not support using a default value - something for which I filed an issue to the fast project (issue [here](https://github.com/microsoft/fast/issues/5116). They submitted a fix for this that became available in the latest version of fast-foundation.

## 👩‍💻 Implementation

Upgraded fast-foundation package and updated text-field implementation to use the new end slot default value.