Sizes as px (#117)
🤨 Rationale

The DSP Style Dictionary is configured to output sizes as rem, but the Nimble designs are currently in px. This change fixes that in the CSS and JS output so we can use these values in design-tokens.ts.

👩‍💻 Implementation

Disabled the scss output (we're not using it)
Leveraged the existing nimble-extensions.js file to override the transform configuration for js and css files.
Updated design-tokens.ts to use token variables for font sizes.
🧪 Testing

Built the DSP interactively and via local build script. Units are correctly updated.

✅ Checklist

 I have updated the project documentation to reflect my changes or determined no changes are needed.