Fix bottom border of text/number field (#113)
## 🤨 Rationale

The bottom border of the text and number fields was growing down instead of up on hover - this resulted in the component shifting up in some cases on hover.

## 👩‍💻 Implementation

Updated the approach to hover border styling used by the select control. Also resolved an issue with the label height that was resulting in the controls being the wrong overall height.