Add tabs component (#118)
🤨 Rationale

Add tabs component with minimal requirements of horizontal tabs with simple text and a disabled state. The according design spec can be found here.

👩‍💻 Implementation

Re-used the tab components from fast-foundation and applied custom styling. We provide a <nimble-tabs> component as the root component, <nimble-tab> to define a tab header button, and a <nimble-tab-panel> to define the content of a tab. The assignment of a <nimble-tab> to its according <nimble-tab-panel> is determined from the order of the components in the DOM. This implementation is provided by fast-foundation.

An example looks like this:

<nimble-tabs>
    <nimble-tab>Tab One</nimble-tab>
    <nimble-tab>Tab Two</nimble-tab>
    <nimble-tab-panel>Tab one content</nimble-tab-panel>
    <nimble-tab-panel>Tab two content</nimble-tab-panel>
</nimble-tabs>

🧪 Testing

Tested it on Chrome, Firefox, Edge, Safari, & Safari Tech Preview.

✅ Checklist

 X - I have updated the project documentation to reflect my changes or determined no changes are needed.