Add tabs toolbar component (#121)
# Pull Request

## 🤨 Rationale

Add `<nimble-tabs-toolbar>` component to add a toolbar with additional components like buttons to `<nimble-tabs>`. The design spec can be found [here](https://xd.adobe.com/view/8ce280ab-1559-4961-945c-182955c7780b-d9b1/screen/b2aa2c0c-03b7-4571-8e0d-de88baf0814b).

### Usage

```html
<nimble-tabs>
    <nimble-tabs-toolbar>
        <nimble-button appearance="ghost">Toolbar Button</nimble-button>
    </nimble-tabs-toolbar>
    <nimble-tab>Tab One</nimble-tab>
    <nimble-tab>Tab Two</nimble-tab>
    <nimble-tab-panel>Tab one content</nimble-tab-panel>
    <nimble-tab-panel>Tab two content</nimble-tab-panel>
</nimble-tabs>
```

## 👩‍💻 Implementation

The tabs component we're using from fast-foundation provides a `end` slot to fill. We add a `toolbar` slot in the end slot that sits right after the tab list and we're using it to add components next to the tab buttons at the top. As a separator, we're using a `<div>` with a border and some margin.

## 🧪 Testing

Relay on storybook.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.