New nimble tree item component (#129)
## 🤨 Rationale

To support the behavior of the navigation tree in SystemLink we need to more directly handle the selected/expanded state of tree items.

 The current behavior looks like the following:
![OldTreeSelectionBehavior](https://user-images.githubusercontent.com/26874831/134586051-fd3448c9-43dd-4c05-9cc3-7c238b8cd4f2.gif)

The new behavior will look like the following, which mirrors the select/expand behavior of the navigation tree in SystemLink, The noted difference is the vertical bar that is present on the selected item, which we will ultimately want to change.
![NewTreeSelectionBehavior](https://user-images.githubusercontent.com/26874831/134586091-01e559c1-7cf9-4a01-a079-ed51b10a2bab.gif)

## 👩‍💻 Implementation

To provide the expected behavior, we simply need to handle the 'selected-change' event from the FAST TreeItem, and set/unset the selected/expanded attributes as needed. Additionally, we are setting our own attribute on TreeItems, to allow our TreeView to manage the current selected item state appropriately when a parent tree item is expanded/collapsed.

## 🧪 Testing

Added automated tests to verify the expected selected/expanded states of tree items based on a click event.