Lighthouse CI Audit (#131)
# Pull Request

## 🤨 Rationale

We don't currently have any automated checks to validate component performance. This PR adds a status check using Lighthouse to generate a set of performance and accessibility metrics that can be used to fail the build or simply review the current performance of a set of test pages that use nimble components.  

## 👩‍💻 Implementation

Followed [this process](https://web.dev/lighthouse-ci/) closely. We're currently pointing Lighthouse to the Angular Example App build directory (`/angular-workspace/dist/example-client-app`), but we can add more pages in the future.
 
## 🧪 Testing

Set initial build errors and warning metrics for performance and accessibility, but this will be an iterative process. Also, need to create the PR before reviewing the status checks work correctly.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.
   - I intend to follow this PR with one that creates a component performance test page in the Example app. I'll update the CONTRIBUTING docs at that time.