Refactor "invalid" attribute on text field to CSS class (#133)
# Pull Request

## 🤨 Rationale

<!---
Provide some background and a description of your work.
What problem does this change solve?

Include links to issues, work items, or other discussions.
-->

When Milan and I were talking about the `invalid` property/attribute on the Nimble text field, we developed some concern that it may be confused with (or maybe clash with) some sort of self-validation behavior in the control. 

## 👩‍💻 Implementation

<!---
Describe how the change addresses the problem. Consider factors such as complexity, alternative solutions, performance impact, etc. 

Consider listing files with important changes or comment on them directly in the pull request.
-->

We considered changing the name of the attribute to something like `showInvalid` to make it clear that it only changes the appearance of the control. But we eventually decided that appearance-only configuration like this probably makes the most sense as a class instead of an attribute.

I removed the text field's class extension which adds the invalid attribute and property. And I updated the styles to match the class instead of the attribute.

## 🧪 Testing

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done on a local build, private CI run results, and additional testing not covered by automatic pull request validation.
-->

Visually checked storybook and Angular example to make sure invalid styles still appear as expected.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.