Icon updates (#137)
🤨 Rationale

Brandon cleaned up the set of Nimble icons, and this PR pulls them in.

👩‍💻 Implementation

The icons are all now laid out on a 16px by 16px canvas, so I needed to update the size of several components… and did so via the ${iconSize} token.
Exported the icons as SVG and removed the fill styling from each. (This will likely not be necessary after the icons are generated as components, but this saved me from adding a number of more specific selectors to apply the fill)
Added an icon control to the Button story to more easily preview icon buttons.
🧪 Testing

Manual Storybook review. I'll also have Brandon review the Chromatic changes, because the controlsArrowExpanderDown16X16 icons could be a bit flatter.

✅ Checklist

 I have updated the project documentation to reflect my changes or determined no changes are needed.