Remove owners from change directory (#144)
# Pull Request

## 🤨 Rationale

I noticed in my PR #143 that @jattasNI and @fredvisser were added as owners only because of the change file. Seems like it doesn't make sense to have particular owners for sub directories if the default owners will be added to every PR anyway.

## 👩‍💻 Implementation

Removed all owners from the `/change` directory. Presumably no one will ever try to make a change which only includes a change file, so we should be covered from owners on other directories.

The last line of [this CODEOWNERS example](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#example-of-a-codeowners-file) seems to imply that clearing owners like this will work.

## 🧪 Testing

None! Hope it works

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.