Switch to the FAST html tagged template for StoryBook (#149)
# Pull Request

## 🤨 Rationale

Before this change the StoryBook stories used string templates to build up HTML. This is different from our components themselves which use the FAST `html` tagged template function. This change updates the stories to use the same `html` function so that stories can benefit from features like syntax highlighting, etc. Another side-benefit is increasing in source examples for leveraging features of FAST.

## 👩‍💻 Implementation

The `createRenderer` function was added that can take the result of the `html` tagged template and initialize a `DOMFragment` with the elements. The `DOMFragment` fulfills the `Story` interface and is passed through the system.

This change also tries to improve the types used for states in the matrix tests to avoid as much direct string interpolation.

## 🧪 Testing

Relying on Chromatic to report render changes. Only intentional change was updating the default value for the Number Field story so that Storybook can create a numeric control to manipulate the value.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed. Didn't see any specific docs that need updating.