Fixed contents of disabled fields shifting on hover  (#150)
## 🤨 Rationale

When working on something else, I noticed that the contents of a disabled select field would shift position on hover. 

![disabled field bug](https://user-images.githubusercontent.com/5454342/136435757-df20c6eb-81fe-4e60-8661-b7dd88edf82f.gif)

The bug also affects the text and number fields.

## 👩‍💻 Implementation

Removed the `padding-bottom` on the main control element when the field is disabled. We add the padding under normal circumstances to compensate for the changing bottom border width of the control on hover. But the disabled control's border does not change on hover, so we should remove the extra padding.

## 🧪 Testing

Verified that the select, text field, and number field appear as expected when disabled and hovered.