Remove support for manual value binding of form controls in Angular (#152)
## 🤨 Rationale

**TL;DR:** We should only support `[(ngModel)]`, `[formControl]`, and `[formControlName]` for Nimble form control value binding in Angular. We will not support binding to value via properties like `[value]` and `(change)` because it requires extra effort and doesn't provide much extra value.

Previously, we aimed to support several different methods of binding to the value of Nimble form controls in Angular:
1. `[(ngModel)]` for two-way binding in Angular template-driven forms or outside of forms
2. `[formControl]` or `[formControlName]` for two-way binding in Angular reactive forms
3. The element's built in properties and events such as `[value]` and `(change)` for manually implementing binding behavior

Angular allows using any of these methods to bind to native `input` elements and Material form controls, and we intended to match that behavior. We are able to support the first and second binding methods by implementing `ControlValueAccessor`s (CVAs) for form controls.

The third binding method is more difficult to support. When not using the CVA (i.e. not using `ngModel`, `formControl`, or `formControlName`), Angular does not allow binding to properties such as `[value]` or `[disabled]` because the compiler does not know that they are properties of the element. To add support for these properties, we would probably need to implement another directive for each form control only for when not using the CVA, using a selector something like this:

`nimble-number-field:not([formControlName]):not([formControl]):not([ngModel])`

This would result in more boilerplate code to write and maintain, and I don't think it adds enough value to make it worthwhile. Milan and I reviewed usages of input elements in Angular apps in the Systemlink repo, and found that the vast majority are using 
`ngModel`, `formControl`, or `formControlName`. The few exceptions were some file inputs (which do use `(change)`), and a few inputs that weren't using any template-based method of getting or setting the input value. For cases where it is desirable or necessary to use a change event instead of two-way binding, consumers can use `(ngModelChange)` binding instead. 

Also, these issues with value binding were masked for a while because of a testing issue. Some of the tests that we've written which exercise value binding via `[value]` and `(change)` were passing, even though the exact same code in an application caused compile errors. The discrepancy seems to come from [the different compilation methods used for ng test vs ng serve and ng build](https://github.com/angular/angular-cli/issues/18177#issuecomment-654796406).

## 👩‍💻 Implementation

Deleted test cases for now unsupported usage. Updated contributing docs. 

## 🧪 Testing

N/A, only deleting tests.