Nimble tree group selection (#160)
## 🤨 Rationale

These changes support the design of the tree where the top-level item will display a vertical bar on its left when one of its leaf tree items is selected.

## 👩‍💻 Implementation

Essentially I've created a new attribute (`group-selected`) to indicate that a particular tree item has a leaf tree item that is currently selected. When selection changes we remove all currently set `group-selected` attributes, and then reset them to the set of tree items that are a parent of the current selected leaf tree item.