Nimble icons (#155)
## 🤨 Rationale

We want to provide clients access to the icons in nimble-tokens via a set of components. This makes it easier, and safer, to consume the icons as desired.

## 👩‍💻 Implementation

There is now a root `Icon` class that all icon components are expected to extend from. Each icon component now has its own folder under the 'icons' folder, and shares the same boilerplate structure. Ultimately, since the implementation is boilerplate, we will be able to automate the creation of each icon component file.

We provide access to theme-defined colors for various states ('pass', 'warning', 'fail', 'regular') via classes. A CSS variable (--icon-color) is also available for a client to override.