Nimble menu icon (#164)
🤨 Rationale

Adding support for icons in menu-items, with consistent indentation on all items in a menu.

👩‍💻 Implementation

Since we decided to use the start slot for icons going forward, this is much simpler than expected. I've added styling to correctly render any svg icon placed in the start slot. I'm also using a grid layout in the menu-items to correctly place the icons and text content in columns, all indented properly.

🧪 Testing

Added icons to both the menu story and menu matrix.

✅ Checklist

X - I have updated the project documentation to reflect my changes or determined no changes are needed.