Nimble menu angular directive (#177)
# Pull Request

## 🤨 Rationale

Using the nimble menu and menu item in Angular apps like the SWIF will be easier with directives that add attribute binding.

## 👩‍💻 Implementation

Added both a directive for nimble-menu and nimble-menu-item.  The menu ended up pretty empty, since it's just a container for menu items.  The items needed a bit more, so they've also got more tests.

## 🧪 Testing

Added spec files to the menu and menu-item directives.  As mentioned, the menu doesn't need much testing.  The menu-item does check that binding works in both directions.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.