Refactor example app (#198)
## 🤨 Rationale

For the web chapter presentation tomorrow, we are planning to demo the example app to show Nimble components in a functional Angular context. To that end, we wanted to refactor the example app to clean it up and better highlight some of our new components (e.g. the drawer).

## 👩‍💻 Implementation

Modified the header component to serve as an app-wide header that contains the menu as well as a "User Settings" menu which exposes the theme select control in a drawer. This more closely adheres to UX patterns in SystemLink while highlighting the new component.

I also refactored the "custom app" such that each component's demo is contained within its own tab (which is itself a demo of the tab control) and the default tab provides some context and links to the Nimble storybook.

Finally, I modified the "logout" button in the header menu to actually route to the login page.