Change to TreeItem to prepare for fast-foundation changes (#211)
## 🤨 Rationale

The fast-foundation TreeItem will likely change in a way that will break our current implementation (see: [https://github.com/microsoft/fast/pull/5350/](https://github.com/microsoft/fast/pull/5350/)). This change will prevent the update to TreeItem from breaking the behavior we desire.

With these changes in place, the update to fast-foundation will resolve the issue ([#194](https://github.com/ni/nimble/issues/194)) where the group of a default selected element in an Angular app doesn't show the proper styling. This change alone, however, will not resolve that without the update, and I'm not sure if there is a way to solve it at this point without it.

## 👩‍💻 Implementation

With the implementation as it was, after the update pointed to above (if and when it occurs), would cause an infinite loop emanating from our handling of the `selected-change` event (in `handleSelectedChange`) where we conditionally update the `selected` attribute to true. This is because the fast-foundation TreeItem would now emit the `selected-change` event whenever the attribute changes, which it wasn't doing before.

Instead, now all we will do is update our `group-selected` attribute in that handler, and manage the `selected` state directly in our click handler. Note that while our behavior will not break after the fast-foundation update happens, we can still improve the implementation after the update, as we will no longer need to explicitly emit the `selected-change` or `expanded-change` events, as these will happen automatically. This is the reason I had to change the test, which was asserting the number of times an event handler was called. With the update, those assertions break with the current implementation, but they aren't really necessary assertions from a behavior perspective, so I've removed them.

## 🧪 Testing

Verified the existing tests work with and without the update to fast-foundation (where the two assertions I removed would have failed with the update to fast-foundation, but they didn't fail against the current version).