Refactor prefers-reduced-motion logic into utility (#241)
## 🤨 Rationale

Resolves #191 

## 👩‍💻 Implementation

Moved creation of media query (`prefers-reduced-motion`) into singleton utility class `PrefersReducedMotionWatcher` in `utilities/style/prefers-reduced-motion.ts`. Clients register a callback that takes a boolean indicating whether animations are disabled. The callback is called when registered and when the value of `prefers-reduced-motion` changes.

The only component currently needing this functionality is the drawer, which I've updated to use the new utility.

## 🧪 Testing

Verified in Storybook that drawer still honors the system's "show animations" setting.

## ✅ Checklist

- [X] I have updated the project documentation to reflect my changes or determined no changes are needed.