Correct styling of icon buttons (#250)
# Pull Request

## 🤨 Rationale

When buttons contain only icons, they should be a square rather than using the padding applied to text buttons.

[XD Spec](https://xd.adobe.com/view/8ce280ab-1559-4961-945c-182955c7780b-d9b1/screen/d022d8af-22f4-4bf2-981c-1dc0c61afece/)

## 👩‍💻 Implementation

Rather than create a new component to be a nimble-icon-button, I leveraged the existing nimble-button. The button handles changes in the default slot by checking to see if any children of the default slot have text content. If not, the `empty-text` class is applied to the control.

Note: This assumes that the text will be in the default slot. If we don't think this is a valid assumption, then we will need to either come up with a different mechanism for automatically determining if a button contains only an icon, or we will need to require a client to inform the button that it contains only an icon.

## 🧪 Testing

Ensured that the existing button matrix tests include the case that the button contains an icon but does not contain any text.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.