Beachball update lock support wildcard dependencies (#252)
# Pull Request

## 🤨 Rationale

When trying to integrate beachball-update-lock I found an existing issue where lock files previously updated by the script would cause dirty changes locally like the following:
![image](https://user-images.githubusercontent.com/1588923/149043314-29c082b8-6ae5-49ea-9ab8-03fee2541a44.png)

The issue is that the [angular-test package uses wildcard dependencies](https://github.com/ni/javascript-styleguide/blob/9616630d4339f362507b0d0c519e79d8acf543de/tests/angular/package.json#L13) for packages within the monorepo. It is a private package that won't be published and only exists within the monorepo. By using wildcard versions beachball can skip updates and dependency update changelogs for the test packages.

The issue is that the beachball-lock-update script did not recognize that the wildcard version in the lock file did not need to be updated and updates the wildcard references to use the new package version number. When npm install runs locally it repairs the lock file and switches back to wildcards based on the package contents and dirties the workspace.

## 👩‍💻 Implementation

- Make beachball-lock-update check for wildcard dependencies
- Modified the angular-workspace package to use wildcard dependencies for packages contained in the monorepo. Should end the [pointless package and changelog dependency updates](https://github.com/ni/nimble/blob/b2a94ccce73b0d7a98ccd148f1304d3ebfcb85e3/angular-workspace/CHANGELOG.md) for the private angular-workspace project.

## 🧪 Testing

Added a test for wildcard deps.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.