Complete button matrix (#256)
# Pull Request

## 🤨 Rationale

I realized that not all button appearances were being tested without a label. This increases the coverage of the button matrix to include all button appearances with (1) text only, (2) icon only, (3) icons & text.

## 👩‍💻 Implementation

Previously the no-text case was treated as an additional appearance that used the default appearance of 'outline'. This meant that only the 'outline' button was ever tested without text.

I modified the test matrix so that no-text is no longer treated as an additional appearance. Instead I combined the valid visibility states into a single input so that we could test all the valid combinations of text/icon while skipping the tests that cases where the button is completely empty.

## 🧪 Testing

This is a test-only change.

## ✅ Checklist

~~- [ ] I have updated the project documentation to reflect my changes or determined no changes are needed.~~
This is a test-only change. There isn't any documentation to be updated.