Create nimble-slider spec (#244)
## 🤨 Rationale

For the upcoming Outlier Detection project, Andrew Dove identified some components which would likely be needed in the UI. Nimble already provides most of these, but there are a few missing, including a numeric slider and a progress bar. I decided to get the ball rolling on implementing these. @jattasNI mentioned to me that there was discussion about a new process for creating new components, including writing a spec document. This PR is a first attempt at a component spec.

## 👩‍💻 Implementation

Created a spec document for the `nimble-slider` and `nimble-slider-label`.
* The document structure is roughly based on the FAST spec documents
* The spec covers API and behaviors, but not visual design (which should be covered by XD mockups)

## 🧪 Testing

N/A