Icon button accessibility (#258)
# Pull Request

## 🤨 Rationale

The icon-only button story had an accessibility violation associated with it. This PR adjusts the way a button becomes 'icon-only' to resolve the accessibility violation.

## 👩‍💻 Implementation

The nimble-button now has an attribute named `content-hidden`. When this attribute is set to `true`, the button behaves as an icon-button regardless of the text content -- the button becomes a square and the content is hidden. The content of the button is used as the button's label in both the case the content is hidden and in the case when the content is not hidden.

## 🧪 Testing

Storybook no longer has an accessibility violation for the icon button story.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.