Create issue templates (#277)
The plan is to start tracking all Nimble work with GitHub issues. We will migrate our Nimble-specific AzDO work items to issues. This PR adds issue templates that should help us create distinct types of issues: Bugs, Features, New Components (a specific kind of Feature), and User Stories.

We can leverage task lists to create a hierarchy of issues, e.g. a feature with user story sub-issues. An item in a task list is not initially backed by a separate issue, but it appears that we can easily do that on-demand.