Ensure text area hover border thickens outward on Safari (#369)
## 🤨 Rationale

On Safari, when the nimble-text-area is focused and hovered, the border thickens inward and is obscured partially by the scrollbar.

## 👩‍💻 Implementation

On Safari, the default styling for `textarea`s when focused sets `outline-offset=-2`. Because I am using an outline to thicken the border on hover, this creates a discrepancy between Safari and other browsers. By using a box-shadow instead of an outline, I avoid this problem.

## 🧪 Testing

Manually tested in Storybook with Chrome and Safari (by using dev tools).

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.