Add styling for primary buttons (#373)
# Pull Request

## 🤨 Rationale
Add styling for primary buttons

Fixes #349

## 👩‍💻 Implementation

Added styling specific to the `nimble-button` for primary buttons when the `primary` CSS class is added to the element. The styling does not affect on a `ghost` button or a `nimble-toggle-button`.

For primary block buttons, Brandon specifically requested that we ensure the border doesn't overlap the background, causing the desired opacity to change. Because of this, I took a different approach between the outline button and block button styling.

## 🧪 Testing
Added 'primary' to button matrix tests -- Note, I added this for the 'Ghost' button as well even though the style doesn't apply there so that we can catch accidentally making primary changes to that appearance.

## ✅ Checklist
- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.