Generating new font tokens. (#390)
🤨 Rationale

We've discovered that there are certain applications that rely on setting sub-font properties (e.g. fontFamily, fontWeight, etc...). So in order to facilitate a re-styling of components to leverage the nimble tokens, in as targeted a way as possible, it is necessary for us to provide tokens for sub-font properties.

👩‍💻 Implementation

I've created a new helper method that allows us to create the set of font tokens related to each of our specific fonts. This method returns an array of CSSDesignToken<string> which we then export. The order of the returned tokens matches the parameter order, which unfortunately can't be enforced.

🧪 Testing

Manual. I modified the Angular example app to use some of the new tokens, which were successfully discovered by type-ahead.

✅ Checklist

 I have updated the project documentation to reflect my changes or determined no changes are needed.